recommendation_two_tower_full_cycle/services/data-ingestion/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
pandas==2.1.3
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
python-json-logger==2.0.7
aiofiles==23.2.1
